fix(tiles): Do not implement Hash
Use `BTree`s instead of `HashMap`s for managing collections of `Hai`s.

Previously, we had the following error:
https://rust-lang.github.io/rust-clippy/master/index.html#derived_hash_with_manual_eq

This is because our implementation of `PartialEq` did not went well with
the implementation of `Hash`.
The following would not have hold:

   k1 == k2 ⇒ hash(k1) == hash(k2)

Indeed, we must ignore aka dora when doing any kind of comparisons
between tiles! The linter was right!