refactor(ai): Simplify function to write an AI
We move the complicated logics with threads in `AiServer::new`.

We can now call `AiServer::new` and simply pass two callback functions
that will handle the decisions of what to do.

In the meantime, we remove the Kakan implementation for the null-bot.
Indeed, the null-bot should not call Kakan. It should just do nothing.